tests: Split large File/Slice repo test to multiple files
