--- internal/pkg/service/buffer/sink/tablesink/storage/repository/file_repo_test.go
@@ -1,2440 +1,0 @@
-package repository_test
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"net/http"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/benbjohnson/clock"
-	"github.com/c2h5oh/datasize"
-	"github.com/jarcoal/httpmock"
-	"github.com/keboola/go-client/pkg/keboola"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.etcd.io/etcd/client/v3/concurrency"
-
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/config"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/column"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/key"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/dependencies"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/statistics"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/repository"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/test"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/volume"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/volume/assignment"
-	commonDeps "github.com/keboola/keboola-as-code/internal/pkg/service/common/dependencies"
-	serviceError "github.com/keboola/keboola-as-code/internal/pkg/service/common/errors"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdop/iterator"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdop/op"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/common/rollback"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/common/utctime"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/etcdhelper"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/etcdlogger"
-)
-
-func TestFileRepository_Operations(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T01:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey1 := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-	sinkKey2 := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-2"}
-	nonExistentFileKey := storage.FileKey{
-		SinkKey: sinkKey1,
-		FileID:  storage.FileID{OpenedAt: utctime.MustParse("2000-01-01T18:00:00.000Z")},
-	}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	rb := rollback.New(d.Logger())
-	client := mocked.TestEtcdClient()
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileRepo := storageRepo.File()
-	sliceRepo := storageRepo.Slice()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 5)
-	}
-
-	// Empty
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		// List - empty
-		files, err := fileRepo.List(projectID).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Empty(t, files)
-		files, err = fileRepo.List(sinkKey1).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Empty(t, files)
-	}
-	{
-		// Get - not found
-		if err := fileRepo.Get(nonExistentFileKey).Do(ctx).Err(); assert.Error(t, err) {
-			assert.Equal(t, `file "123/456/my-source/my-sink-1/2000-01-01T18:00:00.000Z" not found in the sink`, err.Error())
-			serviceError.AssertErrorStatusCode(t, http.StatusNotFound, err)
-		}
-	}
-
-	// Create (the first Rotate) - parent sink doesn't exists
-	// -----------------------------------------------------------------------------------------------------------------
-	// Entity exists only in memory
-	{
-		if err := fileRepo.Rotate(rb, clk.Now(), sinkKey1).Do(ctx).Err(); assert.Error(t, err) {
-			assert.Equal(t, `sink "123/456/my-source/my-sink-1" not found in the source`, err.Error())
-			serviceError.AssertErrorStatusCode(t, http.StatusNotFound, err)
-		}
-	}
-
-	// Create parent branch, source, sinks and tokens
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-
-		source := test.NewSource(sourceKey)
-		require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-
-		sink1 := test.NewSink(sinkKey1)
-		sink1.Table.Storage = sinkStorageConfig(3, []string{"hdd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink1).Do(ctx).Err())
-
-		sink2 := test.NewSink(sinkKey2)
-		sink2.Table.Storage = sinkStorageConfig(3, []string{"ssd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink2).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink1.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink2.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Create (the first Rotate)
-	// See TestFileRepository_Rotate for more rotation tests.
-	// -----------------------------------------------------------------------------------------------------------------
-	var fileKey1, fileKey2 storage.FileKey
-	{
-		// Create 2 files in different sinks
-		clk.Add(time.Hour)
-		file1, err := fileRepo.Rotate(rb, clk.Now(), sinkKey1).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.NotNil(t, file1.StagingStorage.UploadCredentials)
-		fileKey1 = file1.FileKey
-
-		clk.Add(time.Hour)
-		file2, err := fileRepo.Rotate(rb, clk.Now(), sinkKey2).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.NotNil(t, file2.StagingStorage.UploadCredentials)
-		fileKey2 = file2.FileKey
-	}
-	{
-		// List
-		files, err := fileRepo.List(projectID).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Len(t, files, 2)
-		files, err = fileRepo.List(branchKey).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Len(t, files, 2)
-		files, err = fileRepo.List(sourceKey).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Len(t, files, 2)
-		files, err = fileRepo.List(sinkKey1).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Len(t, files, 1)
-		files, err = fileRepo.List(sinkKey2).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Len(t, files, 1)
-	}
-	{
-		// Get
-		result1, err := fileRepo.Get(fileKey1).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, "2000-01-01T02:00:00.000Z", result1.OpenedAt().String())
-		result2, err := fileRepo.Get(fileKey2).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, "2000-01-01T03:00:00.000Z", result2.OpenedAt().String())
-	}
-
-	// File rotation has created slices in assigned volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	var sliceKeys1, sliceKeys2 []storage.SliceKey
-	{
-		// Slices in file1
-		sliceID1 := storage.SliceID{OpenedAt: fileKey1.OpenedAt()}
-		require.NoError(t, sliceRepo.List(fileKey1).Do(ctx).ForEachValue(
-			func(value storage.Slice, header *iterator.Header) error {
-				sliceKeys1 = append(sliceKeys1, value.SliceKey)
-				return nil
-			},
-		))
-		assert.Equal(t, []storage.SliceKey{
-			{SliceID: sliceID1, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey1, VolumeID: "my-volume-1"}}, // hdd
-			{SliceID: sliceID1, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey1, VolumeID: "my-volume-3"}}, // hdd
-			{SliceID: sliceID1, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey1, VolumeID: "my-volume-5"}}, // hdd
-		}, sliceKeys1)
-
-		// Slices in file2
-		sliceID2 := storage.SliceID{OpenedAt: fileKey2.OpenedAt()}
-		require.NoError(t, sliceRepo.List(fileKey2).Do(ctx).ForEachValue(
-			func(value storage.Slice, header *iterator.Header) error {
-				sliceKeys2 = append(sliceKeys2, value.SliceKey)
-				return nil
-			},
-		))
-		assert.Equal(t, []storage.SliceKey{
-			{SliceID: sliceID2, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey2, VolumeID: "my-volume-2"}}, // ssd
-			{SliceID: sliceID2, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey2, VolumeID: "my-volume-4"}}, // ssd
-			{SliceID: sliceID2, FileVolumeKey: storage.FileVolumeKey{FileKey: fileKey2, VolumeID: "my-volume-5"}}, // hdd
-		}, sliceKeys2)
-	}
-
-	// Rotate - already exists
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		if err := fileRepo.Rotate(rb, fileKey1.OpenedAt().Time(), sinkKey1).Do(ctx).Err(); assert.Error(t, err) {
-			assert.Equal(t, `file "123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z" already exists in the sink`, err.Error())
-			serviceError.AssertErrorStatusCode(t, http.StatusConflict, err)
-		}
-	}
-
-	// Increment retry attempt - not found
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		err := fileRepo.IncrementRetry(clk.Now(), nonExistentFileKey, "some error").Do(ctx).Err()
-		if assert.Error(t, err) {
-			assert.Equal(t, `file "123/456/my-source/my-sink-1/2000-01-01T18:00:00.000Z" not found in the sink`, err.Error())
-			serviceError.AssertErrorStatusCode(t, http.StatusNotFound, err)
-		}
-	}
-
-	// Increment retry attempt
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		result, err := fileRepo.IncrementRetry(clk.Now(), fileKey1, "some error").Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, 1, result.RetryAttempt)
-		assert.Equal(t, "some error", result.RetryReason)
-		assert.Equal(t, "2000-01-01T04:00:00.000Z", result.LastFailedAt.String())
-		assert.Equal(t, "2000-01-01T04:02:00.000Z", result.RetryAfter.String())
-
-		file1, err := fileRepo.Get(fileKey1).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, result, file1)
-	}
-
-	// Switch file state to storage.FileClosing
-	// -----------------------------------------------------------------------------------------------------------------
-	switchFileStates(t, ctx, clk, fileRepo, fileKey1, []storage.FileState{
-		storage.FileWriting, storage.FileClosing,
-	})
-
-	// Switch file state - slices are not uploaded
-	// -----------------------------------------------------------------------------------------------------------------
-	if err := fileRepo.StateTransition(clk.Now(), fileKey1, storage.FileClosing, storage.FileImporting).Do(ctx).Err(); assert.Error(t, err) {
-		assert.Equal(t, strings.TrimSpace(`
-unexpected slice "123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z" state:
-- unexpected combination: file state "importing" and slice state "closing"
-`), err.Error())
-		serviceError.AssertErrorStatusCode(t, http.StatusBadRequest, err)
-	}
-
-	// Simulate slices upload to unblock the storage.FileImporting state
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		for _, sliceKey := range sliceKeys1 {
-			switchSliceStates(t, ctx, clk, sliceRepo, sliceKey, []storage.SliceState{
-				storage.SliceClosing, storage.SliceUploading, storage.SliceUploaded,
-			})
-		}
-	}
-
-	// Switch file state to storage.FileImported
-	// -----------------------------------------------------------------------------------------------------------------
-	switchFileStates(t, ctx, clk, fileRepo, fileKey1, []storage.FileState{
-		storage.FileClosing, storage.FileImporting, storage.FileImported,
-	})
-
-	// Switch file state - already in the state
-	// -----------------------------------------------------------------------------------------------------------------
-	if err := fileRepo.StateTransition(clk.Now(), fileKey1, storage.FileImported, storage.FileImported).Do(ctx).Err(); assert.Error(t, err) {
-		assert.Equal(t, `unexpected file "123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z" state transition from "imported" to "imported"`, err.Error())
-		serviceError.AssertErrorStatusCode(t, http.StatusBadRequest, err)
-	}
-
-	// Switch file state - unexpected transition
-	// -----------------------------------------------------------------------------------------------------------------
-	if err := fileRepo.StateTransition(clk.Now(), fileKey1, storage.FileImported, storage.FileImporting).Do(ctx).Err(); assert.Error(t, err) {
-		assert.Equal(t, `unexpected file "123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z" state transition from "imported" to "importing"`, err.Error())
-		serviceError.AssertErrorStatusCode(t, http.StatusBadRequest, err)
-	}
-
-	// Delete
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		assert.NoError(t, fileRepo.Delete(fileKey2).Do(ctx).Err())
-	}
-	{
-		// Get - not found
-		if err := fileRepo.Get(fileKey2).Do(ctx).Err(); assert.Error(t, err) {
-			assert.Equal(t, `file "123/456/my-source/my-sink-2/2000-01-01T03:00:00.000Z" not found in the sink`, err.Error())
-			serviceError.AssertErrorStatusCode(t, http.StatusNotFound, err)
-		}
-	}
-	{
-		// List - empty
-		files, err := fileRepo.List(sinkKey2).Do(ctx).AllKVs()
-		assert.NoError(t, err)
-		assert.Empty(t, files)
-	}
-
-	// Delete - not found
-	// -----------------------------------------------------------------------------------------------------------------
-	if err := fileRepo.Delete(nonExistentFileKey).Do(ctx).Err(); assert.Error(t, err) {
-		assert.Equal(t, `file "123/456/my-source/my-sink-1/2000-01-01T18:00:00.000Z" not found in the sink`, err.Error())
-		serviceError.AssertErrorStatusCode(t, http.StatusNotFound, err)
-	}
-
-	// Check etcd state - file2 has been deleted, but file 1 exists
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, `
-<<<<<
-storage/file/all/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  "closingAt": "2000-01-01T05:00:00.000Z",
-  "importingAt": "2000-01-01T12:00:00.000Z",
-  "importedAt": "2000-01-01T13:00:00.000Z",
-  "columns": [
-    {
-      "type": "datetime",
-      "name": "datetime"
-    },
-    {
-      "type": "body",
-      "name": "body"
-    }
-  ],
-  "assignment": {
-    "config": {
-      "count": 3,
-      "preferredTypes": [
-        "hdd"
-      ]
-    },
-    "volumes": [
-      "my-volume-5",
-      "my-volume-3",
-      "my-volume-1"
-    ]
-  },
-  "local": {
-    "dir": "123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z",
-    "compression": {
-      "type": "gzip",
-      "gzip": {
-        "level": 1,
-        "implementation": "parallel",
-        "blockSize": "256KB",
-        "concurrency": 0
-      }
-    },
-    "diskSync": {
-      "mode": "disk",
-      "wait": true,
-      "checkInterval": 5000000,
-      "countTrigger": 500,
-      "bytesTrigger": "1MB",
-      "intervalTrigger": 50000000
-    },
-    "diskAllocation": {
-      "enabled": true,
-      "size": "100MB",
-      "sizePercent": 110
-    }
-  },
-  "staging": {
-    "compression": {
-      "type": "gzip",
-      "gzip": {
-        "level": 1,
-        "implementation": "parallel",
-        "blockSize": "256KB",
-        "concurrency": 0
-      }
-    },
-    "credentials": {%A},
-    "credentialsExpiration": "2000-01-01T03:00:00.000Z"
-  },
-  "target": {
-    "tableId": "in.bucket.table"
-  }
-}
->>>>>
-
-<<<<<
-storage/file/level/target/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-3/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-3",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-5/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-5",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  %A
-}
->>>>>
-`, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/slice/all|storage/secret/token/|storage/volume"))
-}
-
-func TestFileRepository_Rotate(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T01:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	client := mocked.TestEtcdClient()
-	rb := rollback.New(d.Logger())
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileFacade := storageRepo.File()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 2)
-	}
-
-	// Create parent branch, source, sink and token
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-		source := test.NewSource(sourceKey)
-		require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-		sink := test.NewSink(sinkKey)
-		sink.Table.Storage = sinkStorageConfig(2, []string{"default"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Create (the first file Rotate operation)
-	// -----------------------------------------------------------------------------------------------------------------
-	var file1 storage.File
-	{
-		var err error
-		clk.Add(time.Hour)
-		file1, err = fileFacade.Rotate(rb, clk.Now(), sinkKey).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file1.OpenedAt().Time())
-	}
-
-	// Rotate file (2)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		file2, err := fileFacade.Rotate(rb, clk.Now(), sinkKey).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file2.OpenedAt().Time())
-	}
-
-	// Rotate file (3)
-	// -----------------------------------------------------------------------------------------------------------------
-	var file3 storage.File
-	{
-		var err error
-		clk.Add(time.Hour)
-		file3, err = fileFacade.Rotate(rb, clk.Now(), sinkKey).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file3.OpenedAt().Time())
-	}
-
-	// Check Storage API calls
-	// -----------------------------------------------------------------------------------------------------------------
-	// File prepare endpoint should be called N times
-	assert.Equal(t, 3, transport.GetCallCountInfo()["POST /v2/storage/branch/456/files/prepare"])
-	assert.Equal(t, 0, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Test rollback, delete file endpoint should be called N times
-	rb.Invoke(ctx)
-	assert.Equal(t, 3, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Check etcd state
-	//   - Only the last file is in the storage.FileWriting state.
-	//   - Other files are in the storage.FileClosing state.
-	//   - Slices are switched to the storage.SliceClosing state together with the file state transition.
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, `
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T03:00:00.000Z",
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T03:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T03:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T04:00:00.000Z",
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T04:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T04:00:00.000Z",
-  "type": "csv",
-  "state": "writing",
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T03:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-2/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-2",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T03:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T03:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T03:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T04:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T03:00:00.000Z",
-  "volumeId": "my-volume-2",
-  "sliceOpenedAt": "2000-01-01T03:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T04:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-1/2000-01-01T04:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T04:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T04:00:00.000Z",
-  "type": "csv",
-  "state": "writing",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-2/2000-01-01T04:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T04:00:00.000Z",
-  "volumeId": "my-volume-2",
-  "sliceOpenedAt": "2000-01-01T04:00:00.000Z",
-  "type": "csv",
-  "state": "writing",
-  %A
-}
->>>>>
-`, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all/|storage/secret/token/|storage/volume/"))
-}
-
-func TestFileRepository_Rotate_FileResourceError(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T01:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	client := mocked.TestEtcdClient()
-	rb := rollback.New(d.Logger())
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileFacade := storageRepo.File()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	transport.RegisterResponder(
-		http.MethodPost,
-		fmt.Sprintf("/v2/storage/branch/%d/files/prepare", branchKey.BranchID),
-		func(request *http.Request) (*http.Response, error) {
-			response, err := httpmock.NewJsonResponse(http.StatusUnauthorized, keboola.StorageError{
-				Message: "some error",
-			})
-			response.Request = request // required by the StorageError
-			return response, err
-		},
-	)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 1)
-	}
-
-	// Create parent branch, source, sink and token
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-		source := test.NewSource(sourceKey)
-		require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-		sink := test.NewSink(sinkKey)
-		sink.Table.Storage = sinkStorageConfig(1, []string{"default"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Create (the first file Rotate operation)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		_, err := fileFacade.Rotate(rb, clk.Now(), sinkKey).Do(ctx).ResultOrErr()
-		if assert.Error(t, err) {
-			assert.Equal(t, strings.TrimSpace(`
-cannot create file resource:
-- some error, method: "POST", url: "https://connection.keboola.local/v2/storage/branch/456/files/prepare", httpCode: "401"
-`), err.Error())
-		}
-	}
-
-	// Check Storage API calls
-	// -----------------------------------------------------------------------------------------------------------------
-	assert.Equal(t, 1, transport.GetCallCountInfo()["POST /v2/storage/branch/456/files/prepare"])
-
-	// Check etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, "", etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all/|storage/secret/token/|storage/volume/"))
-}
-
-func TestFileRepository_RotateOnSinkMod(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T02:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	client := mocked.TestEtcdClient()
-	rb := rollback.New(d.Logger())
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileFacade := storageRepo.File()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 1)
-	}
-
-	// Create parent branch, source and sink
-	// -----------------------------------------------------------------------------------------------------------------
-	var sink definition.Sink
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-		source := test.NewSource(sourceKey)
-		require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-		sink = test.NewSink(sinkKey)
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Create (the first file Rotate)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		sink.Table.Mapping.Columns = column.Columns{column.Body{Name: "body1"}}
-		file1, err := fileFacade.RotateOnSinkMod(rb, clk.Now(), sink).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file1.OpenedAt().Time())
-	}
-
-	// Rotate file (2)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		sink.Table.Mapping.Columns = column.Columns{column.Body{Name: "body2"}}
-		file2, err := fileFacade.RotateOnSinkMod(rb, clk.Now(), sink).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file2.OpenedAt().Time())
-	}
-
-	// Rotate file (3)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		sink.Table.Mapping.Columns = column.Columns{column.Body{Name: "body3"}}
-		file3, err := fileFacade.RotateOnSinkMod(rb, clk.Now(), sink).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file3.OpenedAt().Time())
-	}
-
-	// Check Storage API calls
-	// -----------------------------------------------------------------------------------------------------------------
-	// File prepare endpoint should be called N times
-	assert.Equal(t, 3, transport.GetCallCountInfo()["POST /v2/storage/branch/456/files/prepare"])
-	assert.Equal(t, 0, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Test rollback, delete file endpoint should be called N times
-	rb.Invoke(ctx)
-	assert.Equal(t, 3, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Check etcd state
-	//   - Only the last file is in the storage.FileWriting state.
-	//   - Other files are in the storage.FileClosing state.
-	//   - Columns definition is copied from the provided Sink entity.
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, `
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T03:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T03:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T04:00:00.000Z",
-  "columns": [
-    {
-      "type": "body",
-      "name": "body1"
-    }
-  ],
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T04:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T04:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T05:00:00.000Z",
-  "columns": [
-    {
-      "type": "body",
-      "name": "body2"
-    }
-  ],
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T05:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T05:00:00.000Z",
-  "type": "csv",
-  "state": "writing",
-  "columns": [
-    {
-      "type": "body",
-      "name": "body3"
-    }
-  ],
-  %A
-  "staging": {
-    %A
-    "credentials": {
-      "id": 1%d%d%d,
-      %A
-    },
-    %A
-  },
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T03:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T03:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T04:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-1/2000-01-01T04:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T04:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T04:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T05:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T05:00:00.000Z/my-volume-1/2000-01-01T05:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T05:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T05:00:00.000Z",
-  "type": "csv",
-  "state": "writing",
-  %A
-}
->>>>>
-`, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all|storage/secret/token/|storage/volume/writer/my-volume-1"))
-}
-
-func TestFileRepository_CloseAllIn(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T01:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	rb := rollback.New(d.Logger())
-	client := mocked.TestEtcdClient()
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileRepo := storageRepo.File()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Log etcd operations
-	var etcdLogs bytes.Buffer
-	rawClient := d.EtcdClient()
-	rawClient.KV = etcdlogger.KVLogWrapper(rawClient.KV, &etcdLogs, etcdlogger.WithMinimal())
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 1)
-	}
-
-	// Create sink
-	// -----------------------------------------------------------------------------------------------------------------
-	branch := test.NewBranch(branchKey)
-	require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-	source := test.NewSource(sourceKey)
-	require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-	sink := test.NewSink(sinkKey)
-	require.NoError(t, defRepo.Sink().Create("Create sink", &sink).Do(ctx).Err())
-	require.NoError(t, tokenRepo.Put(sink.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-
-	// Create 2 files, with 2 slices
-	// -----------------------------------------------------------------------------------------------------------------
-	require.NoError(t, fileRepo.Rotate(rb, clk.Now(), sinkKey).Do(ctx).Err())
-	clk.Add(time.Hour)
-	require.NoError(t, fileRepo.Rotate(rb, clk.Now(), sinkKey).Do(ctx).Err())
-
-	// Close the last file
-	// -----------------------------------------------------------------------------------------------------------------
-	clk.Add(time.Hour)
-	etcdLogs.Reset()
-	require.NoError(t, fileRepo.CloseAllIn(clk.Now(), sinkKey).Do(ctx).Err())
-	closeAllInEtcdLogs := etcdLogs.String()
-
-	// Check etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdlogger.Assert(t, `
-// CloseAllIn - AtomicOp - Read Phase
-➡️  TXN
-  ➡️  THEN:
-  // Sink
-  001 ➡️  GET "definition/sink/active/123/456/my-source/my-sink-1"
-  // Local files
-  002 ➡️  GET ["storage/file/level/local/123/456/my-source/my-sink-1/", "storage/file/level/local/123/456/my-source/my-sink-10")
-  // Local slices
-  003 ➡️  GET ["storage/slice/level/local/123/456/my-source/my-sink-1/", "storage/slice/level/local/123/456/my-source/my-sink-10")
-✔️  TXN | succeeded: true
-
-// CloseAllIn - AtomicOp - Write Phase
-➡️  TXN
-  ➡️  IF:
-  // Sink
-  001 "definition/sink/active/123/456/my-source/my-sink-1" MOD GREATER 0
-  002 "definition/sink/active/123/456/my-source/my-sink-1" MOD LESS %d
-  // Local files
-  003 ["storage/file/level/local/123/456/my-source/my-sink-1/", "storage/file/level/local/123/456/my-source/my-sink-10") MOD GREATER 0
-  004 ["storage/file/level/local/123/456/my-source/my-sink-1/", "storage/file/level/local/123/456/my-source/my-sink-10") MOD LESS %d
-  005 "storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  006 "storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z" MOD GREATER 0
-  // Local slices
-  007 ["storage/slice/level/local/123/456/my-source/my-sink-1/", "storage/slice/level/local/123/456/my-source/my-sink-10") MOD GREATER 0
-  008 ["storage/slice/level/local/123/456/my-source/my-sink-1/", "storage/slice/level/local/123/456/my-source/my-sink-10") MOD LESS %d
-  009 "storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  010 "storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z" MOD GREATER 0
-  ➡️  THEN:
-  // Save closed file - in two copies, in "all" and <level> prefix
-  001 ➡️  PUT "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z"
-  002 ➡️  PUT "storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z"
-  // Save closed slice - in two copies, in "all" and <level> prefix
-  003 ➡️  PUT "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z"
-  004 ➡️  PUT "storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z"
-✔️  TXN | succeeded: true
-`, closeAllInEtcdLogs)
-
-	// Check etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	expectedEtcdState := `
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z
------
-{
-  %A
-  "fileOpenedAt": "2000-01-01T01:00:00.000Z",
-  %A
-  "state": "closing",
-  "closingAt": "2000-01-01T02:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z
------
-{
-  %A
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  %A
-  "state": "closing",
-  "closingAt": "2000-01-01T03:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T01:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T01:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T02:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source/my-sink-1/2000-01-01T02:00:00.000Z/my-volume-1/2000-01-01T02:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T02:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T02:00:00.000Z",
-  "type": "csv",
-  "state": "closing",
-  "closingAt": "2000-01-01T03:00:00.000Z",
-  %A
-}
->>>>>
-`
-	etcdhelper.AssertKVsString(t, client, expectedEtcdState, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all/|storage/secret/token/|storage/volume/"))
-
-	// Call CloseAllIn again - no change
-	clk.Add(time.Hour)
-	require.NoError(t, fileRepo.CloseAllIn(clk.Now(), sinkKey).Do(ctx).Err())
-	etcdhelper.AssertKVsString(t, client, expectedEtcdState, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all/|storage/secret/token/|storage/volume/"))
-}
-
-func TestFileRepository_RotateAllIn(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T02:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey1 := key.SourceKey{BranchKey: branchKey, SourceID: "my-source-1"}
-	sourceKey2 := key.SourceKey{BranchKey: branchKey, SourceID: "my-source-2"}
-	sinkKey1 := key.SinkKey{SourceKey: sourceKey1, SinkID: "my-sink-1"}
-	sinkKey2 := key.SinkKey{SourceKey: sourceKey1, SinkID: "my-sink-2"}
-	sinkKey3 := key.SinkKey{SourceKey: sourceKey1, SinkID: "my-sink-3"}
-	sinkKey4 := key.SinkKey{SourceKey: sourceKey2, SinkID: "my-sink-4"}
-	sinkKey5 := key.SinkKey{SourceKey: sourceKey2, SinkID: "my-sink-5"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	client := mocked.TestEtcdClient()
-	rb := rollback.New(d.Logger())
-	defRepo := d.DefinitionRepository()
-	storageRepo := d.StorageRepository()
-	fileFacade := storageRepo.File()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Log etcd operations
-	var etcdLogs bytes.Buffer
-	rawClient := d.EtcdClient()
-	rawClient.KV = etcdlogger.KVLogWrapper(rawClient.KV, &etcdLogs, etcdlogger.WithMinimal())
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Create parent branch, source, sinks and tokens
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-		source1 := test.NewSource(sourceKey1)
-		require.NoError(t, defRepo.Source().Create("Create source", &source1).Do(ctx).Err())
-		source2 := test.NewSource(sourceKey2)
-		require.NoError(t, defRepo.Source().Create("Create source", &source2).Do(ctx).Err())
-		sink1 := test.NewSink(sinkKey1)
-		sink1.Table.Storage = sinkStorageConfig(3, []string{"ssd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink1).Do(ctx).Err())
-		sink2 := test.NewSink(sinkKey2)
-		sink2.Table.Storage = sinkStorageConfig(3, []string{"hdd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink2).Do(ctx).Err())
-		sink3 := test.NewSink(sinkKey3)
-		sink3.Table.Storage = sinkStorageConfig(2, []string{"ssd", "hdd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink3).Do(ctx).Err())
-		sink4 := test.NewSink(sinkKey4)
-		sink4.Table.Storage = sinkStorageConfig(1, []string{"ssd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink4).Do(ctx).Err())
-		sink5 := test.NewSink(sinkKey5)
-		sink5.Table.Storage = sinkStorageConfig(1, []string{"hdd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink5).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink1.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink2.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink3.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink4.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink5.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 5)
-	}
-
-	// Create (the first file Rotate)
-	// -----------------------------------------------------------------------------------------------------------------
-	var rotateAllInEtcdLogs string
-	{
-		clk.Add(time.Hour)
-		etcdLogs.Reset()
-		require.NoError(t, fileFacade.RotateAllIn(rb, clk.Now(), branchKey).Do(ctx).Err())
-		rotateAllInEtcdLogs = etcdLogs.String()
-	}
-
-	// Rotate file (2)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		require.NoError(t, fileFacade.RotateAllIn(rb, clk.Now(), branchKey).Do(ctx).Err())
-	}
-
-	// Rotate file (3)
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		require.NoError(t, fileFacade.RotateAllIn(rb, clk.Now(), branchKey).Do(ctx).Err())
-	}
-
-	// Check Storage API calls
-	// -----------------------------------------------------------------------------------------------------------------
-	// File prepare endpoint should be called N times
-	assert.Equal(t, 15, transport.GetCallCountInfo()["POST /v2/storage/branch/456/files/prepare"])
-	assert.Equal(t, 0, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Test rollback, delete file endpoint should be called N times
-	rb.Invoke(ctx)
-	assert.Equal(t, 15, transport.GetCallCountInfo()[`DELETE =~/v2/storage/branch/456/files/\d+$`])
-
-	// Check etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdlogger.Assert(t, `
-// RotateAllIn - AtomicOp - Read Phase
-➡️  TXN
-  ➡️  THEN:
-  // Sinks
-  001 ➡️  GET ["definition/sink/active/123/456/", "definition/sink/active/123/4560")
-  // Volumes
-  002 ➡️  GET ["storage/volume/writer/", "storage/volume/writer0")
-  // Local files
-  003 ➡️  GET ["storage/file/level/local/123/456/", "storage/file/level/local/123/4560")
-  // Local slices
-  004 ➡️  GET ["storage/slice/level/local/123/456/", "storage/slice/level/local/123/4560")
-✔️  TXN | succeeded: true
-
-// RotateAllIn - FileResourcesProvider - Get tokens
-➡️  TXN
-  ➡️  THEN:
-  001 ➡️  GET "storage/secret/token/123/456/my-source-1/my-sink-1"
-  002 ➡️  GET "storage/secret/token/123/456/my-source-1/my-sink-2"
-  003 ➡️  GET "storage/secret/token/123/456/my-source-1/my-sink-3"
-  004 ➡️  GET "storage/secret/token/123/456/my-source-2/my-sink-4"
-  005 ➡️  GET "storage/secret/token/123/456/my-source-2/my-sink-5"
-✔️  TXN | succeeded: true
-
-// RotateAllIn - MaxUsedDiskSizeBySliceIn
-➡️  TXN
-  ➡️  THEN:
-  001 ➡️  GET ["storage/stats/staging/123/456/my-source-1/my-sink-1/", "storage/stats/staging/123/456/my-source-1/my-sink-10")
-  002 ➡️  GET ["storage/stats/target/123/456/my-source-1/my-sink-1/", "storage/stats/target/123/456/my-source-1/my-sink-10")
-  003 ➡️  GET ["storage/stats/staging/123/456/my-source-1/my-sink-2/", "storage/stats/staging/123/456/my-source-1/my-sink-20")
-  004 ➡️  GET ["storage/stats/target/123/456/my-source-1/my-sink-2/", "storage/stats/target/123/456/my-source-1/my-sink-20")
-  005 ➡️  GET ["storage/stats/staging/123/456/my-source-1/my-sink-3/", "storage/stats/staging/123/456/my-source-1/my-sink-30")
-  006 ➡️  GET ["storage/stats/target/123/456/my-source-1/my-sink-3/", "storage/stats/target/123/456/my-source-1/my-sink-30")
-  007 ➡️  GET ["storage/stats/staging/123/456/my-source-2/my-sink-4/", "storage/stats/staging/123/456/my-source-2/my-sink-40")
-  008 ➡️  GET ["storage/stats/target/123/456/my-source-2/my-sink-4/", "storage/stats/target/123/456/my-source-2/my-sink-40")
-  009 ➡️  GET ["storage/stats/staging/123/456/my-source-2/my-sink-5/", "storage/stats/staging/123/456/my-source-2/my-sink-50")
-  010 ➡️  GET ["storage/stats/target/123/456/my-source-2/my-sink-5/", "storage/stats/target/123/456/my-source-2/my-sink-50")
-✔️  TXN | succeeded: true
-
-// RotateAllIn - AtomicOp - Write Phase
-➡️  TXN
-  ➡️  IF:
-  // Sinks
-  001 ["definition/sink/active/123/456/", "definition/sink/active/123/4560") MOD GREATER 0
-  002 ["definition/sink/active/123/456/", "definition/sink/active/123/4560") MOD LESS %d
-  003 "definition/sink/active/123/456/my-source-1/my-sink-1" MOD GREATER 0
-  004 "definition/sink/active/123/456/my-source-1/my-sink-2" MOD GREATER 0
-  005 "definition/sink/active/123/456/my-source-1/my-sink-3" MOD GREATER 0
-  006 "definition/sink/active/123/456/my-source-2/my-sink-4" MOD GREATER 0
-  007 "definition/sink/active/123/456/my-source-2/my-sink-5" MOD GREATER 0
-  // Volumes
-  008 ["storage/volume/writer/", "storage/volume/writer0") MOD GREATER 0
-  009 ["storage/volume/writer/", "storage/volume/writer0") MOD LESS %d
-  010 "storage/volume/writer/my-volume-1" MOD GREATER 0
-  011 "storage/volume/writer/my-volume-2" MOD GREATER 0
-  012 "storage/volume/writer/my-volume-3" MOD GREATER 0
-  013 "storage/volume/writer/my-volume-4" MOD GREATER 0
-  014 "storage/volume/writer/my-volume-5" MOD GREATER 0
-  // Local files
-  015 ["storage/file/level/local/123/456/", "storage/file/level/local/123/4560") MOD EQUAL 0
-  // Local slices
-  016 ["storage/slice/level/local/123/456/", "storage/slice/level/local/123/4560") MOD EQUAL 0
-  // New objects must not exists
-  017 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  018 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  019 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  020 "storage/file/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  021 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  022 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  023 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  024 "storage/file/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  025 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  026 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  027 "storage/file/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  028 "storage/slice/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  029 "storage/file/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  030 "storage/slice/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  031 "storage/file/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  ➡️  THEN:
-  // Save closed and opened files and slices - in two copies, in "all" and <level> prefix
-  001 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  002 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  003 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z"
-  004 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z"
-  005 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  006 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  007 ➡️  PUT "storage/file/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z"
-  008 ➡️  PUT "storage/file/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z"
-  009 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  010 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  011 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z"
-  012 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z"
-  013 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z"
-  014 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z"
-  015 ➡️  PUT "storage/file/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z"
-  016 ➡️  PUT "storage/file/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z"
-  017 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  018 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  019 ➡️  PUT "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z"
-  020 ➡️  PUT "storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z"
-  021 ➡️  PUT "storage/file/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z"
-  022 ➡️  PUT "storage/file/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z"
-  023 ➡️  PUT "storage/slice/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  024 ➡️  PUT "storage/slice/level/local/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z"
-  025 ➡️  PUT "storage/file/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z"
-  026 ➡️  PUT "storage/file/level/local/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z"
-  027 ➡️  PUT "storage/slice/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  028 ➡️  PUT "storage/slice/level/local/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z"
-  029 ➡️  PUT "storage/file/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z"
-  030 ➡️  PUT "storage/file/level/local/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z"
-  // IF conditions in the ELSE branch are used to detect a specific cause of failure, for example, "file already exists"
-  ➡️  ELSE:
-  001 ➡️  TXN
-  001   ➡️  IF:
-  001   001 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   004 "storage/file/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   005 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   006 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   007 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   008 "storage/file/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   009 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   010 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   011 "storage/file/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   012 "storage/slice/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   013 "storage/file/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   014 "storage/slice/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   015 "storage/file/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   ➡️  ELSE:
-  001   001 ➡️  TXN
-  001   001   ➡️  IF:
-  001   001   001 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   002 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   003 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   004 "storage/file/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   ➡️  ELSE:
-  001   001   001 ➡️  TXN
-  001   001   001   ➡️  IF:
-  001   001   001   001 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   002 ➡️  TXN
-  001   001   002   ➡️  IF:
-  001   001   002   001 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   003 ➡️  TXN
-  001   001   003   ➡️  IF:
-  001   001   003   001 "storage/slice/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   001   004 ➡️  TXN
-  001   001   004   ➡️  IF:
-  001   001   004   001 "storage/file/all/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002 ➡️  TXN
-  001   002   ➡️  IF:
-  001   002   001 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   002 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   003 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   004 "storage/file/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   ➡️  ELSE:
-  001   002   001 ➡️  TXN
-  001   002   001   ➡️  IF:
-  001   002   001   001 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   002 ➡️  TXN
-  001   002   002   ➡️  IF:
-  001   002   002   001 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   003 ➡️  TXN
-  001   002   003   ➡️  IF:
-  001   002   003   001 "storage/slice/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   002   004 ➡️  TXN
-  001   002   004   ➡️  IF:
-  001   002   004   001 "storage/file/all/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003 ➡️  TXN
-  001   003   ➡️  IF:
-  001   003   001 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003   002 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003   003 "storage/file/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003   ➡️  ELSE:
-  001   003   001 ➡️  TXN
-  001   003   001   ➡️  IF:
-  001   003   001   001 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003   002 ➡️  TXN
-  001   003   002   ➡️  IF:
-  001   003   002   001 "storage/slice/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   003   003 ➡️  TXN
-  001   003   003   ➡️  IF:
-  001   003   003   001 "storage/file/all/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   004 ➡️  TXN
-  001   004   ➡️  IF:
-  001   004   001 "storage/slice/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   004   002 "storage/file/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   004   ➡️  ELSE:
-  001   004   001 ➡️  TXN
-  001   004   001   ➡️  IF:
-  001   004   001   001 "storage/slice/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   004   002 ➡️  TXN
-  001   004   002   ➡️  IF:
-  001   004   002   001 "storage/file/all/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   005 ➡️  TXN
-  001   005   ➡️  IF:
-  001   005   001 "storage/slice/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   005   002 "storage/file/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   005   ➡️  ELSE:
-  001   005   001 ➡️  TXN
-  001   005   001   ➡️  IF:
-  001   005   001   001 "storage/slice/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-  001   005   002 ➡️  TXN
-  001   005   002   ➡️  IF:
-  001   005   002   001 "storage/file/all/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z" MOD EQUAL 0
-
-✔️  TXN | succeeded: true
-`, rotateAllInEtcdLogs)
-
-	// Check etcd state
-	//   - Only the last file per the Sink is in the storage.FileWriting state.
-	//   - Other files per the Sink are in the storage.FileClosing state.
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, `
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-1/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-1/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-2/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-2/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-3/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-1/my-sink-3/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-4/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-4/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-5/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/file/level/local/123/456/my-source-2/my-sink-5/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-2/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-3/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T04:00:00.000Z/my-volume-4/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T05:00:00.000Z/my-volume-2/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T05:00:00.000Z/my-volume-4/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-1/2000-01-01T05:00:00.000Z/my-volume-5/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-1/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-3/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T04:00:00.000Z/my-volume-1/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T04:00:00.000Z/my-volume-3/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T04:00:00.000Z/my-volume-5/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T05:00:00.000Z/my-volume-1/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T05:00:00.000Z/my-volume-3/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-2/2000-01-01T05:00:00.000Z/my-volume-5/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T03:00:00.000Z/my-volume-4/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T04:00:00.000Z/my-volume-2/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T04:00:00.000Z/my-volume-4/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T05:00:00.000Z/my-volume-2/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-1/my-sink-3/2000-01-01T05:00:00.000Z/my-volume-4/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-4/2000-01-01T03:00:00.000Z/my-volume-2/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-4/2000-01-01T04:00:00.000Z/my-volume-4/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-4/2000-01-01T05:00:00.000Z/my-volume-2/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-5/2000-01-01T03:00:00.000Z/my-volume-5/2000-01-01T03:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-5/2000-01-01T04:00:00.000Z/my-volume-3/2000-01-01T04:00:00.000Z
------
-{
-%A
-  "state": "closing",
-%A
-}
->>>>>
-
-<<<<<
-storage/slice/level/local/123/456/my-source-2/my-sink-5/2000-01-01T05:00:00.000Z/my-volume-5/2000-01-01T05:00:00.000Z
------
-{
-%A
-  "state": "writing",
-%A
-}
->>>>>
-`, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all|storage/secret/token/|storage/volume/"))
-}
-
-func TestFileRepository_StateTransition(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	clk := clock.NewMock()
-	clk.Set(utctime.MustParse("2000-01-01T00:00:00.000Z").Time())
-
-	// Fixtures
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-
-	// Get services
-	d, mocked := dependencies.NewMockedTableSinkScope(t, config.New(), commonDeps.WithClock(clk))
-	client := mocked.TestEtcdClient()
-	rb := rollback.New(d.Logger())
-	defRepo := d.DefinitionRepository()
-	statsRepo := d.StatisticsRepository()
-	storageRepo := d.StorageRepository()
-	fileRepo := storageRepo.File()
-	sliceRepo := storageRepo.Slice()
-	tokenRepo := storageRepo.Token()
-	volumeRepo := storageRepo.Volume()
-
-	// Log etcd operations
-	var etcdLogs bytes.Buffer
-	rawClient := d.EtcdClient()
-	rawClient.KV = etcdlogger.KVLogWrapper(rawClient.KV, &etcdLogs, etcdlogger.WithMinimal())
-
-	// Mock file API calls
-	transport := mocked.MockedHTTPTransport()
-	mockStorageAPICalls(t, clk, branchKey, transport)
-
-	// Register active volumes
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		session, err := concurrency.NewSession(client)
-		require.NoError(t, err)
-		defer func() { require.NoError(t, session.Close()) }()
-		registerWriterVolumes(t, ctx, volumeRepo, session, 2)
-	}
-
-	// Create parent branch, source, sink and token
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		branch := test.NewBranch(branchKey)
-		require.NoError(t, defRepo.Branch().Create(&branch).Do(ctx).Err())
-		source := test.NewSource(sourceKey)
-		require.NoError(t, defRepo.Source().Create("Create source", &source).Do(ctx).Err())
-		sink := test.NewSink(sinkKey)
-		sink.Table.Storage = sinkStorageConfig(2, []string{"ssd"})
-		require.NoError(t, defRepo.Sink().Create("Create sink", &sink).Do(ctx).Err())
-		require.NoError(t, tokenRepo.Put(sink.SinkKey, keboola.Token{Token: "my-token"}).Do(ctx).Err())
-	}
-
-	// Create (the first file Rotate)
-	// -----------------------------------------------------------------------------------------------------------------
-	var file storage.File
-	{
-		clk.Add(time.Hour)
-		var err error
-		file, err = fileRepo.Rotate(rb, clk.Now(), sinkKey).Do(ctx).ResultOrErr()
-		require.NoError(t, err)
-		assert.Equal(t, clk.Now(), file.OpenedAt().Time())
-	}
-
-	// Get file slices
-	// -----------------------------------------------------------------------------------------------------------------
-	var sliceKey1, sliceKey2 storage.SliceKey
-	{
-		slices, err := sliceRepo.List(file.FileKey).Do(ctx).All()
-		require.NoError(t, err)
-		require.Len(t, slices, 2)
-		sliceKey1 = slices[0].SliceKey
-		sliceKey2 = slices[1].SliceKey
-	}
-
-	// Put slice statistics values - they should be moved with the file state transitions
-	// -----------------------------------------------------------------------------------------------------------------
-	require.NoError(t, statsRepo.Put(ctx, []statistics.PerSlice{
-		{
-			SliceKey: sliceKey1,
-			Value: statistics.Value{
-				SlicesCount:      1,
-				FirstRecordAt:    sliceKey1.OpenedAt(),
-				LastRecordAt:     sliceKey1.OpenedAt().Add(time.Minute),
-				RecordsCount:     12,
-				UncompressedSize: 10 * datasize.MB,
-				CompressedSize:   10 * datasize.MB,
-			},
-		},
-		{
-			SliceKey: sliceKey2,
-			Value: statistics.Value{
-				SlicesCount:      1,
-				FirstRecordAt:    sliceKey2.OpenedAt(),
-				LastRecordAt:     sliceKey2.OpenedAt().Add(time.Minute),
-				RecordsCount:     34,
-				UncompressedSize: 20 * datasize.MB,
-				CompressedSize:   20 * datasize.MB,
-			},
-		},
-	}))
-
-	// Switch file to the storage.FileClosing state by StateTransition, it is not possible
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		err := fileRepo.StateTransition(clk.Now(), file.FileKey, storage.FileWriting, storage.FileClosing).Do(ctx).Err()
-		if assert.Error(t, err) {
-			assert.Equal(t, `unexpected file transition to the state "closing", use Rotate* or Close* methods`, err.Error())
-		}
-	}
-
-	// Switch file to the storage.FileClosing state by CloseAllIn
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		require.NoError(t, fileRepo.CloseAllIn(clk.Now(), sinkKey).Do(ctx).Err())
-	}
-
-	// Both slices are uploaded
-	// -----------------------------------------------------------------------------------------------------------------
-	{
-		clk.Add(time.Hour)
-		require.NoError(t, sliceRepo.StateTransition(clk.Now(), sliceKey1, storage.SliceClosing, storage.SliceUploading).Do(ctx).Err())
-		require.NoError(t, sliceRepo.StateTransition(clk.Now(), sliceKey1, storage.SliceUploading, storage.SliceUploaded).Do(ctx).Err())
-
-		clk.Add(time.Hour)
-		require.NoError(t, sliceRepo.StateTransition(clk.Now(), sliceKey2, storage.SliceClosing, storage.SliceUploading).Do(ctx).Err())
-		require.NoError(t, sliceRepo.StateTransition(clk.Now(), sliceKey2, storage.SliceUploading, storage.SliceUploaded).Do(ctx).Err())
-	}
-
-	// Switch file to the storage.FileImporting state
-	// -----------------------------------------------------------------------------------------------------------------
-	var toImportingEtcdLogs string
-	{
-		etcdLogs.Reset()
-		clk.Add(time.Hour)
-		require.NoError(t, fileRepo.StateTransition(clk.Now(), file.FileKey, storage.FileClosing, storage.FileImporting).Do(ctx).Err())
-		toImportingEtcdLogs = etcdLogs.String()
-	}
-
-	// Switch file to the storage.FileImported state
-	// -----------------------------------------------------------------------------------------------------------------
-	var toImportedEtcdLogs string
-	{
-		etcdLogs.Reset()
-		clk.Add(time.Hour)
-		require.NoError(t, fileRepo.StateTransition(clk.Now(), file.FileKey, storage.FileImporting, storage.FileImported).Do(ctx).Err())
-		toImportedEtcdLogs = etcdLogs.String()
-	}
-
-	// Check etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdlogger.Assert(t, `
-// File StateTransition - AtomicOp - Read Phase
-➡️  TXN
-  ➡️  THEN:
-  // File
-  001 ➡️  GET "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  // All slices in file
-  002 ➡️  GET ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0")
-  // Local statistics
-  003 ➡️  GET ["storage/stats/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/stats/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0")
-✔️  TXN | succeeded: true
-
-// File StateTransition - AtomicOp - Write Phase
-➡️  TXN
-  ➡️  IF:
-  // File
-  001 "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  002 "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z" MOD LESS %d
-  // All slices in file
-  003 ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD GREATER 0
-  004 ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD LESS %d
-  005 "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  006 "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  // Local statistics
-  007 ["storage/stats/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/stats/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD EQUAL 0
-  ➡️  THEN:
-  // Save modified file
-  001 ➡️  PUT "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  002 ➡️  PUT "storage/file/level/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  003 ➡️  DEL "storage/file/level/local/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-✔️  TXN | succeeded: true
-`, toImportingEtcdLogs)
-	etcdlogger.Assert(t, `
-// File StateTransition - AtomicOp - Read Phase
-➡️  TXN
-  ➡️  THEN:
-  // File
-  001 ➡️  GET "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  // All slices in file
-  002 ➡️  GET ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0")
-  // Staging statistics
-  003 ➡️  GET ["storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0")
-✔️  TXN | succeeded: true
-
-// File StateTransition - AtomicOp - Write Phase
-➡️  TXN
-  ➡️  IF:
-  // File
-  001 "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  002 "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z" MOD LESS %d
-  // All slices in file
-  003 ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD GREATER 0
-  004 ["storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD LESS %d
-  005 "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  006 "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z" MOD GREATER 0
-  // Staging statistics
-  007 ["storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD GREATER 0
-  008 ["storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/", "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z0") MOD LESS %d
-  009 "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z/value" MOD GREATER 0
-  010 "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z/value" MOD GREATER 0
-  ➡️  THEN:
-  // Save modified file and slices - in two copies, in "all" and <level> prefix
-  001 ➡️  PUT "storage/file/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  002 ➡️  PUT "storage/file/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  003 ➡️  DEL "storage/file/level/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z"
-  004 ➡️  PUT "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z"
-  005 ➡️  PUT "storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z"
-  006 ➡️  DEL "storage/slice/level/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z"
-  007 ➡️  PUT "storage/slice/all/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z"
-  008 ➡️  PUT "storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z"
-  009 ➡️  DEL "storage/slice/level/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z"
-  // Move file statistics
-  010 ➡️  PUT "storage/stats/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z/value"
-  011 ➡️  DEL "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z/value"
-  012 ➡️  PUT "storage/stats/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z/value"
-  013 ➡️  DEL "storage/stats/staging/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z/value"
-✔️  TXN | succeeded: true
-`, toImportedEtcdLogs)
-
-	// Check final etcd state
-	// -----------------------------------------------------------------------------------------------------------------
-	etcdhelper.AssertKVsString(t, client, `
-<<<<<
-storage/file/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T01:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  "closingAt": "2000-01-01T02:00:00.000Z",
-  "importingAt": "2000-01-01T05:00:00.000Z",
-  "importedAt": "2000-01-01T06:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T01:00:00.000Z",
-  "volumeId": "my-volume-1",
-  "sliceOpenedAt": "2000-01-01T01:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  "closingAt": "2000-01-01T02:00:00.000Z",
-  "uploadingAt": "2000-01-01T03:00:00.000Z",
-  "uploadedAt": "2000-01-01T03:00:00.000Z",
-  "importedAt": "2000-01-01T06:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z
------
-{
-  "projectId": 123,
-  "branchId": 456,
-  "sourceId": "my-source",
-  "sinkId": "my-sink-1",
-  "fileOpenedAt": "2000-01-01T01:00:00.000Z",
-  "volumeId": "my-volume-2",
-  "sliceOpenedAt": "2000-01-01T01:00:00.000Z",
-  "type": "csv",
-  "state": "imported",
-  "closingAt": "2000-01-01T02:00:00.000Z",
-  "uploadingAt": "2000-01-01T04:00:00.000Z",
-  "uploadedAt": "2000-01-01T04:00:00.000Z",
-  "importedAt": "2000-01-01T06:00:00.000Z",
-  %A
-}
->>>>>
-
-<<<<<
-storage/stats/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-1/2000-01-01T01:00:00.000Z/value
------
-{
-  "slicesCount": 1,
-  "firstRecordAt": "2000-01-01T01:00:00.000Z",
-  "lastRecordAt": "2000-01-01T01:01:00.000Z",
-  "recordsCount": 12,
-  "uncompressedSize": "10MB",
-  "compressedSize": "10MB"
-}
->>>>>
-
-<<<<<
-storage/stats/target/123/456/my-source/my-sink-1/2000-01-01T01:00:00.000Z/my-volume-2/2000-01-01T01:00:00.000Z/value
------
-{
-  "slicesCount": 1,
-  "firstRecordAt": "2000-01-01T01:00:00.000Z",
-  "lastRecordAt": "2000-01-01T01:01:00.000Z",
-  "recordsCount": 34,
-  "uncompressedSize": "20MB",
-  "compressedSize": "20MB"
-}
->>>>>
-`, etcdhelper.WithIgnoredKeyPattern("^definition/|storage/file/all/|storage/slice/all|storage/secret/token/|storage/volume/"))
-}
-
-func registerWriterVolumes(t *testing.T, ctx context.Context, volumeRepo *repository.VolumeRepository, session *concurrency.Session, count int) {
-	t.Helper()
-	volumes := []volume.Metadata{
-		{
-			VolumeID: "my-volume-1",
-			Spec:     volume.Spec{NodeID: "node-a", Type: "hdd", Label: "1", Path: "hdd/1"},
-		},
-		{
-			VolumeID: "my-volume-2",
-			Spec:     volume.Spec{NodeID: "node-b", Type: "ssd", Label: "2", Path: "ssd/2"},
-		},
-		{
-			VolumeID: "my-volume-3",
-			Spec:     volume.Spec{NodeID: "node-b", Type: "hdd", Label: "3", Path: "hdd/3"},
-		},
-		{
-			VolumeID: "my-volume-4",
-			Spec:     volume.Spec{NodeID: "node-b", Type: "ssd", Label: "4", Path: "ssd/4"},
-		},
-		{
-			VolumeID: "my-volume-5",
-			Spec:     volume.Spec{NodeID: "node-c", Type: "hdd", Label: "5", Path: "hdd/5"},
-		},
-	}
-
-	if count < 1 || count > 5 {
-		panic(errors.New("count must be 1-5"))
-	}
-
-	txn := op.Txn(session.Client())
-	for _, vol := range volumes[:count] {
-		txn.Merge(volumeRepo.RegisterWriterVolume(vol, session.Lease()))
-	}
-	require.NoError(t, txn.Do(ctx).Err())
-}
-
-func sinkStorageConfig(count int, preferred []string) *storage.ConfigPatch {
-	return &storage.ConfigPatch{
-		VolumeAssignment: &assignment.Config{
-			Count:          count,
-			PreferredTypes: preferred,
-		},
-	}
-}
-
-func switchFileStates(t *testing.T, ctx context.Context, clk *clock.Mock, fileRepo *repository.FileRepository, fileKey storage.FileKey, states []storage.FileState) {
-	t.Helper()
-	from := states[0]
-	for _, to := range states[1:] {
-		clk.Add(time.Hour)
-
-		// File must be closed by the CloseAllIn method
-		var file storage.File
-		var err error
-		if to == storage.FileClosing {
-			require.Equal(t, storage.FileWriting, from)
-			require.NoError(t, fileRepo.CloseAllIn(clk.Now(), fileKey.SinkKey).Do(ctx).Err())
-			file, err = fileRepo.Get(fileKey).Do(ctx).ResultOrErr()
-			require.NoError(t, err)
-		} else {
-			file, err = fileRepo.StateTransition(clk.Now(), fileKey, from, to).Do(ctx).ResultOrErr()
-			require.NoError(t, err)
-		}
-
-		// File state has been switched
-		assert.Equal(t, to, file.State)
-
-		// Retry should be reset
-		assert.Equal(t, 0, file.RetryAttempt)
-		assert.Nil(t, file.LastFailedAt)
-
-		// Check timestamp
-		switch to {
-		case storage.FileClosing:
-			assert.Equal(t, utctime.From(clk.Now()).String(), file.ClosingAt.String())
-		case storage.FileImporting:
-			assert.Equal(t, utctime.From(clk.Now()).String(), file.ImportingAt.String())
-		case storage.FileImported:
-			assert.Equal(t, utctime.From(clk.Now()).String(), file.ImportedAt.String())
-		default:
-			panic(errors.Errorf(`unexpected file state "%s"`, to))
-		}
-
-		from = to
-	}
-}
